[PyTorch Edge][type] Add type support for NamedTuple custom class (export)
Add type support for namedtule custom class. For the namedtuple type, it will deserailize to the following format
```
('type_name',
    ('NamedTuple,
        ('filed_name_1', 'field_type_1')
        ('filed_name_2', 'field_type_2')
    )
)
```
It will be a forward incompatible change. However this type is never supported and exported before and we don't have a proper way to backport it. The optimum solution to ship this change is probably
1. Update the change for import without the change to export. So the runtime can read the new format, but no new format will be exported.
2. Update the change to export the new type. So runtime can export new format.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/62612


ghstack-source-id: 135655056

Differential Revision: [D30004716](https://our.internmc.facebook.com/intern/diff/D30004716/)